historian: Rename the parser module to gossipd
We were getting a couple of path/import conflicts, renaming just to be
safe.
